scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "beautifulsoup4>=4.13.5",
    "boto3>=1.35.0",
    "click>=8.3.0",
    "lxml>=6.0.1",
    "openai>=1.108.1",
    "psycopg2-binary>=2.9.10",
    "pybloom-live>=4.0.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "rq>=2.6.0",
    "scrapy>=2.11",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
